Move difficulty adjustment mods to a separate class
