+++ osu.Game.Rulesets.Osu/Mods/OsuDifficultyAdjustmentMods.cs
@@ -1,0 +1,18 @@
+﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
+
+using osu.Game.Rulesets.Mods;
+
+namespace osu.Game.Rulesets.Osu.Mods
+{
+    public class OsuDifficultyAdjustmentMods : DifficultyAdjustmentMods
+    {
+        protected override Mod[] Mods => new Mod[]
+        {
+            new OsuModDoubleTime(),
+            new OsuModHalfTime(),
+            new OsuModEasy(),
+            new OsuModHardRock(),
+        };
+    }
+}
