+++ osu.Game.Rulesets.Taiko/Difficulty/TaikoDifficultyCalculator.cs
@@ -6,7 +6,6 @@ using System.Collections.Generic;
 using osu.Game.Beatmaps;
 using osu.Game.Rulesets.Difficulty;
 using osu.Game.Rulesets.Mods;
-using osu.Game.Rulesets.Taiko.Mods;
 using osu.Game.Rulesets.Taiko.Objects;
 
 namespace osu.Game.Rulesets.Taiko.Difficulty
@@ -62,14 +61,6 @@ namespace osu.Game.Rulesets.Taiko.Difficulty
 
             return starRating;
         }
-
-        protected override Mod[] DifficultyAdjustmentMods => new Mod[]
-        {
-            new TaikoModDoubleTime(),
-            new TaikoModHalfTime(),
-            new TaikoModEasy(),
-            new TaikoModHardRock(),
-        };
 
         private bool calculateStrainValues()
         {
