+++ osu.Game.Rulesets.Taiko/Mods/TaikoDifficultyAdjustmentMods.cs
@@ -1,0 +1,18 @@
+﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
+
+using osu.Game.Rulesets.Mods;
+
+namespace osu.Game.Rulesets.Taiko.Mods
+{
+    public class TaikoDifficultyAdjustmentMods : DifficultyAdjustmentMods
+    {
+        protected override Mod[] Mods => new Mod[]
+        {
+            new TaikoModDoubleTime(),
+            new TaikoModHalfTime(),
+            new TaikoModEasy(),
+            new TaikoModHardRock(),
+        };
+    }
+}
