+++ osu.Game/Rulesets/Difficulty/DifficultyCalculator.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using osu.Framework.Extensions.IEnumerableExtensions;
@@ -36,44 +35,6 @@ namespace osu.Game.Rulesets.Difficulty
         protected virtual void PreprocessHitObjects()
         {
         }
-
-        /// <summary>
-        /// Creates all <see cref="Mod"/> combinations which adjust the <see cref="Beatmap"/> difficulty.
-        /// </summary>
-        public Mod[] CreateDifficultyAdjustmentModCombinations()
-        {
-            return createDifficultyAdjustmentModCombinations(Enumerable.Empty<Mod>(), DifficultyAdjustmentMods).ToArray();
-
-            IEnumerable<Mod> createDifficultyAdjustmentModCombinations(IEnumerable<Mod> currentSet, Mod[] adjustmentSet, int currentSetCount = 0, int adjustmentSetStart = 0)
-            {
-                // Initial-case: Empty current set
-                if (currentSetCount == 0)
-                    yield return new NoModMod();
-
-                if (currentSetCount == 1)
-                    yield return currentSet.Single();
-
-                if (currentSetCount > 1)
-                    yield return new MultiMod(currentSet.ToArray());
-
-                // Apply mods in the adjustment set recursively. Using the entire adjustment set would result in duplicate multi-mod mod
-                // combinations in further recursions, so a moving subset is used to eliminate this effect
-                for (int i = adjustmentSetStart; i < adjustmentSet.Length; i++)
-                {
-                    var adjustmentMod = adjustmentSet[i];
-                    if (currentSet.Any(c => c.IncompatibleMods.Any(m => m.IsInstanceOfType(adjustmentMod))))
-                        continue;
-
-                    foreach (var combo in createDifficultyAdjustmentModCombinations(currentSet.Append(adjustmentMod), adjustmentSet, currentSetCount + 1, i + 1))
-                        yield return combo;
-                }
-            }
-        }
-
-        /// <summary>
-        /// Retrieves all <see cref="Mod"/>s which adjust the <see cref="Beatmap"/> difficulty.
-        /// </summary>
-        protected virtual Mod[] DifficultyAdjustmentMods => Array.Empty<Mod>();
 
         public abstract double Calculate(Dictionary<string, double> categoryDifficulty = null);
     }
